Refactor delivery cost calculation to the delivery calculator service
